+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 * updated for `php` 7.1 and `phpunit` 8.
 
 ## 1.2 branch
+### 1.2.5
+* added `can_be_string()` global function;
+* added `is_positive_or_fail()` global function and the
+    `NotPositiveException` exception class.
+
 ### 1.2.4
 * fixed bug for `is_url()` global function, it correctly recognizes the url with
     brackets.